is_it_a_cat.html:  More descriptive/fun display of confidence.
Done by cursor + claude 3.7.
This was the prompt:

Can we now make it a little more fun?  Add some descriptive text about the confidence of the prediction.  Eg if just one label is above some high threshold - eg 97 percent, say "it's definitely a {label}".   If the confidences are very close and/or both low, say "I have no idea".   If one is significantly higher than the other but below the high level, say "it's probably a {label}".     Please think of other possible cases.  Be creative and funny.  I don't know what typical confidence ranges are, but they usually seem to be either in the high 90s or very low.  You decide!